+++ src/main/java/com/elca/jobfairmanagementsystem/service/impl/ExperienceServiceImpl.java
@@ -17,7 +17,6 @@ import java.util.stream.Collectors;
 @Service
 @Transactional
 public class ExperienceServiceImpl implements ExperienceService {
-
     private final ExperienceMapper experienceMapper;
     private final ExperienceRepository experienceRepo;
 
@@ -27,20 +26,15 @@ public class ExperienceServiceImpl implements ExperienceService {
     }
 
     @Override
-    public List<ExperienceDto> searchAllExperience() throws ExperienceNotFoundException {
-
+    public List<ExperienceDto> findAllExperience() throws ExperienceNotFoundException {
         List<Experience> experiences = experienceRepo.findAll();
-
         if (experiences.size() != 0) {
-
             return experiences.stream()
                     .map(experienceMapper::experienceEntityToDto)
                     .collect(Collectors.toList());
         } else
-
             throw new ExperienceNotFoundException(ErrorMessages.NO_EXPERIENCE_AVAILABLE.toString());
     }
-
 
     @Override
     public void saveExperience(ExperienceDto experienceDto) {
@@ -48,41 +42,33 @@ public class ExperienceServiceImpl implements ExperienceService {
         experienceRepo.save(experience);
     }
 
-
     @Override
-    public void deleteExperience(Long experienceId) {
-
-        experienceRepo.deleteById(experienceId);
+    public void deleteExperience(Long experienceId) throws ExperienceNotFoundException {
+        ExperienceDto experience = findByExperienceId(experienceId);
+        if (experience != null) {
+            experienceRepo.deleteById(experienceId);
+        } else {
+            throw new ExperienceNotFoundException(ErrorMessages.EXPERIENCE_NOT_FOUND.toString());
+        }
     }
 
     @Override
     public void updateExperience(ExperienceDto experienceDto) throws ExperienceNotFoundException {
-
         ExperienceDto experience = findByExperienceId(experienceDto.getExperienceId());
-
         if (experience != null) {
-
             experience.setCandidateId(experienceDto.getCandidateId());
             experience.setCompanyName(experienceDto.getCompanyName());
             experience.setDuration(experienceDto.getDuration());
             experience.setPosition(experienceDto.getPosition());
             experienceRepo.save(experienceMapper.experienceDtoToEntity(experience));
         } else
-
             throw new ExperienceNotFoundException(ErrorMessages.EXPERIENCE_NOT_FOUND.toString());
-
     }
 
     @Override
     public ExperienceDto findByExperienceId(Long experienceId) throws ExperienceNotFoundException {
-
         Optional<Experience> optionalExperience = experienceRepo.findById(experienceId);
-
-        var experience = optionalExperience.orElseThrow(() ->
-
-                new ExperienceNotFoundException(ErrorMessages.EXPERIENCE_NOT_FOUND.toString()));
-
+        var experience = optionalExperience.orElseThrow(() -> new ExperienceNotFoundException(ErrorMessages.EXPERIENCE_NOT_FOUND.toString()));
         return experienceMapper.experienceEntityToDto(experience);
     }
-
 }
