+++ src/main/java/com/elca/jobfairmanagementsystem/service/impl/SkillServiceImpl.java
@@ -2,6 +2,8 @@ package com.elca.jobfairmanagementsystem.service.impl;
 
 import com.elca.jobfairmanagementsystem.dto.SkillDto;
 import com.elca.jobfairmanagementsystem.entity.Skill;
+import com.elca.jobfairmanagementsystem.exception.ErrorMessages;
+import com.elca.jobfairmanagementsystem.exception.SkillNotFoundException;
 import com.elca.jobfairmanagementsystem.mapper.SkillMapper;
 import com.elca.jobfairmanagementsystem.repository.SkillRepository;
 import com.elca.jobfairmanagementsystem.service.SkillService;
@@ -21,46 +23,52 @@ public class SkillServiceImpl implements SkillService {
     private final SkillRepository skillRepository;
     private final SkillMapper skillMapper;
 
-    public SkillServiceImpl(SkillRepository skillRepository, SkillMapper skillMapper)
-    {
-        this.skillRepository=skillRepository;
-        this.skillMapper= skillMapper;
+    public SkillServiceImpl(SkillRepository skillRepository, SkillMapper skillMapper) {
+        this.skillRepository = skillRepository;
+        this.skillMapper = skillMapper;
     }
-    @Override
-    public List<SkillDto> searchAllSkills() {
-        List<Skill> skills = skillRepository.findAll();
 
+    @Override
+    public List<SkillDto> findAllSkills() throws SkillNotFoundException {
+        List<Skill> skills = skillRepository.findAll();
         if (skills.size() != 0) {
             return skills.stream().map(skillMapper::skillEntityToDto).collect(Collectors.toList());
-        }
-        else{
-            return null;
+        } else {
+            throw new SkillNotFoundException(ErrorMessages.NO_SKILL_AVAILABLE.toString());
         }
     }
 
     @Override
-    public SkillDto findSkillById(long skillId) {
-        Optional<Skill> optionalSkill =skillRepository.findById(skillId);
-        var skill = optionalSkill.orElse(null);
+    public SkillDto findSkillById(long skillId) throws SkillNotFoundException {
+        Optional<Skill> optionalSkill = skillRepository.findById(skillId);
+        var skill = optionalSkill.orElseThrow(() -> new SkillNotFoundException(ErrorMessages.SKILL_NOT_FOUND.toString()));
         return skillMapper.skillEntityToDto(skill);
     }
 
-
     @Override
-    public Skill saveSkill(SkillDto skillDto) {
+    public void saveSkill(SkillDto skillDto) {
         Skill newSkill = skillMapper.skillDtoToEntity(skillDto);
-        return skillRepository.save(newSkill);
+        skillRepository.save(newSkill);
     }
 
     @Override
-    public void deleteSkill(Long skillId) {
-        skillRepository.deleteById(skillId);
+    public void deleteSkill(Long skillId) throws SkillNotFoundException {
+        var getSkill = findSkillById(skillId);
+        if (getSkill != null) {
+            skillRepository.deleteById(skillId);
+        } else {
+            throw new SkillNotFoundException(ErrorMessages.SKILL_NOT_FOUND.toString());
+        }
     }
 
     @Override
-    public void updateSkill(SkillDto skillDto) {
+    public void updateSkill(SkillDto skillDto) throws SkillNotFoundException {
         var updateSkill = findSkillById(skillDto.getSkillId());
-        updateSkill.setSkillName(skillDto.getSkillName());
-        skillRepository.save(skillMapper.skillDtoToEntity(updateSkill));
+        if (updateSkill != null) {
+            updateSkill.setSkillName(skillDto.getSkillName());
+            skillRepository.save(skillMapper.skillDtoToEntity(updateSkill));
+        } else {
+            throw new SkillNotFoundException(ErrorMessages.SKILL_NOT_FOUND.toString());
+        }
     }
 }
